remove store stats
