+++ proxy_tests/proxy/normal.rs
@@ -20,7 +20,7 @@ mod store {
         for id in cluster.engines.keys() {
             let engine = cluster.get_tiflash_engine(*id);
             assert_eq!(
-                engine.ffi_hub.as_ref().unwrap().get_store_stats().capacity,
+                engine.get_store_stats().fs_stats.capacity_size,
                 444444
             );
         }
